# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Foodie API
  version: 1.0.0
  description: |
    Peer-to-peer homemade-food marketplace API. Customers browse dishes from
    home kitchens, place pickup orders, and leave reviews; food makers manage
    their kitchen, menu, orders, and earnings; admins oversee the platform.
    Money is in integer cents. Auth is a Bearer JWT from /api/auth/*.
servers:
  - url: http://localhost:4000
    description: Local dev
security: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    OrderStatus:
      type: string
      enum: [NEW, ACTIVE, COMPLETED, CANCELLED]
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
        details: { type: array, items: { type: object } }
...
    Kitchen:
      type: object
      properties:
        id: { type: string }
        ownerId: { type: string }
        name: { type: string }
        about: { type: string }
        location: { type: string }
        avatarUrl: { type: [string, 'null'] }
        bannerUrl: { type: [string, 'null'] }
        verified: { type: boolean }
        rating: { type: number }
        reviewCount: { type: integer }
        followerCount: { type: integer }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
    Dish:
      type: object
      properties:
        id: { type: string }
        kitchenId: { type: string }
        name: { type: string }
        description: { type: string }
        priceCents: { type: integer, example: 1400 }
        category: { type: string, example: indian }
        tags: { type: string, description: 'comma-separated', example: 'Veg, Spicy' }
        portion: { type: string, example: '1 portion = 350g' }
        quantityTotal: { type: integer }
        quantityLeft: { type: integer }
        pickupStart: { type: string }
        pickupEnd: { type: string }
        imageUrl: { type: [string, 'null'] }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
    OrderItem:
      type: object
      properties:
        id: { type: string }
        dishId: { type: string }
        nameSnapshot: { type: string }
        priceCentsSnapshot: { type: integer }
        qty: { type: integer }
    Order:
      type: object
      properties:
        id: { type: string, example: FD123456 }
        customerId: { type: string }
        kitchenId: { type: string }
        status: { $ref: '#/components/schemas/OrderStatus' }
        subtotalCents: { type: integer }
        serviceFeeCents: { type: integer }
        totalCents: { type: integer }
        paymentMethod: { type: string }
        pickupSlot: { type: string }
        paymentIntentId: { type: [string, 'null'] }
        createdAt: { type: string, format: date-time }
        items: { type: array, items: { $ref: '#/components/schemas/OrderItem' } }
...
    Payout:
      type: object
      properties:
        id: { type: string }
        kitchenId: { type: string }
        period: { type: string, example: '2026-09' }
        grossCents: { type: integer }
        commissionCents: { type: integer }
        tipsCents: { type: integer }
        netCents: { type: integer }
        status: { type: string, enum: [PENDING, PAID] }
        createdAt: { type: string, format: date-time }
    Paginated:
      type: object
      required: [data, page, limit, total]
      properties:
        data: { type: array, items: { type: object } }
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
paths:
...
  /api/dishes:
    get:
      tags: [dishes]
      summary: List active dishes with filters and pagination
      parameters:
        - in: query
          name: category
          schema: { type: string }
          example: indian
        - in: query
          name: search
          schema: { type: string }
          example: biryani
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Paginated'
                  - type: object
                    properties:
                      data: { type: array, items: { $ref: '#/components/schemas/Dish' } }
  /api/dishes/{id}:
    get:
      tags: [dishes]
      summary: Get one dish with its kitchen
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Dish' } } } }
        '404': { description: Not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
  /api/kitchens/{id}:
    get:
      tags: [kitchens]
      summary: Kitchen profile with menu and recent reviews
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Kitchen' } } } }
        '404': { description: Not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /api/kitchens/{id}/dishes:
    get:
      tags: [kitchens]
      summary: Active dishes for a kitchen (today's menu)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Dish' } } } } }
        '404': { description: Not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
  /api/orders:
    post:
      tags: [orders]
      summary: Place an order (all items must be from one kitchen)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [items, paymentMethod, pickupSlot]
              properties:
                items:
                  type: array
                  minItems: 1
                  items:
                    type: object
                    required: [dishId, qty]
                    properties:
                      dishId: { type: string, example: chicken-biryani }
                      qty: { type: integer, minimum: 1, maximum: 20, example: 2 }
                paymentMethod: { type: string, example: 'Apple Pay' }
                pickupSlot: { type: string, example: '5:00 PM - 7:00 PM' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  order: { $ref: '#/components/schemas/Order' }
                  paymentIntent: { type: object, description: 'Mocked until Stripe is wired' }
        '400': { description: Mixed kitchens, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '404': { description: Dish not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '409': { description: Insufficient stock / unavailable, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
  /api/orders/{id}:
    get:
      tags: [orders]
      summary: Order detail (customer, kitchen owner, or admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Order' } } } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '403': { description: Forbidden, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '404': { description: Not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
  /api/maker/kitchen:
    get:
      tags: [maker]
      summary: The signed-in maker's kitchen (404 if none yet)
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Kitchen' } } } }
        '404': { description: No kitchen yet, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /api/maker/kitchens:
    post:
      tags: [maker]
      summary: Create your kitchen (one per maker, MAKER role)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, example: "Priya's Kitchen" }
                about: { type: string }
                location: { type: string }
                avatarUrl: { type: string, format: uri }
                bannerUrl: { type: string, format: uri }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Kitchen' } } } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '403': { description: Requires MAKER role, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '409': { description: Kitchen already exists, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /api/maker/kitchens/{id}:
    patch:
      tags: [maker]
      summary: Edit your kitchen
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                about: { type: string }
                location: { type: string }
                avatarUrl: { type: string, format: uri }
                bannerUrl: { type: string, format: uri }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Kitchen' } } } }
        '403': { description: Not your kitchen, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '404': { description: Not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /api/maker/dishes:
    get:
      tags: [maker]
      summary: All dishes for your kitchen (including inactive)
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Dish' } } } } }
    post:
      tags: [maker]
      summary: List a new dish on your menu
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, priceCents, category, quantityTotal]
              properties:
                name: { type: string, example: 'Chicken Biryani' }
                description: { type: string }
                priceCents: { type: integer, minimum: 1, example: 1400 }
                category: { type: string, example: indian }
                tags: { type: string, example: 'Non-veg, Medium spice' }
                portion: { type: string }
                quantityTotal: { type: integer, minimum: 0, example: 10 }
                pickupStart: { type: string, example: '5:00 PM' }
                pickupEnd: { type: string, example: '7:00 PM' }
                imageUrl: { type: string, format: uri }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Dish' } } } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '403': { description: Requires MAKER role, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '404': { description: No kitchen yet, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /api/maker/dishes/{id}:
    patch:
      tags: [maker]
      summary: Edit a dish, including the isActive toggle
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                description: { type: string }
                priceCents: { type: integer, minimum: 1 }
                category: { type: string }
                tags: { type: string }
                portion: { type: string }
                quantityTotal: { type: integer, minimum: 0 }
                pickupStart: { type: string }
                pickupEnd: { type: string }
                imageUrl: { type: string, format: uri }
                isActive: { type: boolean }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Dish' } } } }
        '404': { description: Not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /api/maker/orders:
    get:
      tags: [maker]
      summary: Incoming orders for your kitchen
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: status
          schema: { $ref: '#/components/schemas/OrderStatus' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Order' } } } } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '403': { description: Requires MAKER role, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /api/maker/orders/{id}:
    patch:
      tags: [maker]
      summary: Accept (NEW→ACTIVE) or reject (NEW→CANCELLED, stock restored) an order
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action]
              properties:
                action: { type: string, enum: [accept, reject] }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Order' } } } }
        '404': { description: Not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '409': { description: Order is not NEW, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /api/maker/earnings:
    get:
      tags: [maker]
      summary: Earnings summary, 12-week series, and payouts
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  summary:
                    type: object
                    properties:
                      grossCents: { type: integer }
                      commissionCents: { type: integer }
                      tipsCents: { type: integer }
                      netCents: { type: integer }
                      commissionRate: { type: number, example: 0.1 }
                  weekly:
                    type: array
                    items:
                      type: object
                      properties:
                        week: { type: string, example: '2026-09-21' }
                        grossCents: { type: integer }
                  payouts: { type: array, items: { $ref: '#/components/schemas/Payout' } }
  /api/maker/dashboard:
    get:
      tags: [maker]
      summary: Today's sales, open order count, food remaining
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  todaySalesCents: { type: integer }
                  orderCount: { type: integer }
                  foodRemaining: { type: integer }
...